requirements.txt
aiogram==2.25.1
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0